requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
httpx==0.27.0
PyYAML==6.0.1
cryptography==42.0.5
paho-mqtt==1.6.1
w1thermsensor==2.0.0
